# internal/spicegen/commerce/artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Commerce API",
    "version": "0.1.0-dev"
  },
  "paths": {
    "/catalog": {
      "get": {
        "operationId": "Controller_Catalog",
        "summary": "Catalog",
        "tags": [
          "github.com/spice-framework/commerce/orders"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse_0106bdd3d0"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "x-spice-symbol": "spice:symbol:v1|method|42:github.com/spice-framework/commerce/orders|10:Controller|7:Catalog",
        "x-spice-module": "github.com/spice-framework/commerce/orders"
      }
    },
    "/orders": {
      "post": {
        "operationId": "Controller_Place",
        "summary": "Place",
        "tags": [
          "github.com/spice-framework/commerce/orders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderBody_31dfed54f6"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse_40f077d876"
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          },
          "403": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "SpicePrincipal": []
          }
        ],
        "x-spice-symbol": "spice:symbol:v1|method|42:github.com/spice-framework/commerce/orders|10:Controller|5:Place",
        "x-spice-module": "github.com/spice-framework/commerce/orders",
        "x-spice-authorization": {
          "owner": "github.com/spice-framework/commerce/orders",
          "allScopes": [
            "orders:write"
          ]
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "Controller_Get",
        "summary": "Get",
        "tags": [
          "github.com/spice-framework/commerce/orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse_40f077d876"
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          },
          "403": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "SpicePrincipal": []
          }
        ],
        "x-spice-symbol": "spice:symbol:v1|method|42:github.com/spice-framework/commerce/orders|10:Controller|3:Get",
        "x-spice-module": "github.com/spice-framework/commerce/orders",
        "x-spice-authorization": {
          "owner": "github.com/spice-framework/commerce/orders",
          "allScopes": [
            "orders:read"
          ],
          "expression": "authenticated \u0026\u0026 hasScope(\"orders:read\")"
        }
      }
    },
    "/orders/{id}/receipt": {
      "post": {
        "operationId": "Controller_SendReceipt",
        "summary": "SendReceipt",
        "tags": [
          "github.com/spice-framework/commerce/orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse_c5fb030a4c"
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          },
          "403": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          },
          "default": {
            "description": "RFC 9457 problem response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/SpiceProblem"
                }
              }
            }
          }
        },
        "security": [
          {
            "SpicePrincipal": []
          }
        ],
        "x-spice-symbol": "spice:symbol:v1|method|42:github.com/spice-framework/commerce/orders|10:Controller|11:SendReceipt",
        "x-spice-module": "github.com/spice-framework/commerce/orders",
        "x-spice-authorization": {
          "owner": "github.com/spice-framework/commerce/orders",
          "allScopes": [
            "orders:notify"
          ]
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CatalogResponse_0106bdd3d0": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "unit_price_cents": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "sku",
          "unit_price_cents"
        ]
      },
      "OrderResponse_40f077d876": {
        "type": "object",
        "properties": {
          "authorization_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int64"
          },
          "sku": {
            "type": "string"
          },
          "total_cents": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "authorization_id",
          "id",
          "quantity",
          "sku",
          "total_cents"
        ]
      },
      "PlaceOrderBody_31dfed54f6": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "quantity"
        ]
      },
      "ReceiptResponse_c5fb030a4c": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "attachment": {
            "type": "string"
          },
          "message_id": {
            "type": "string"
          },
          "transport": {
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "attachment",
          "message_id",
          "transport"
        ]
      },
      "SpiceProblem": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "format": "uri-reference"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      }
    },
    "securitySchemes": {
      "SpicePrincipal": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}